Enable Windows build (#474)
* define M_PI because windows doesn't

* package management for windows

* updated package detection for windows

* added definition of M_PI because windows does not have it

* added windows.h

* changed posix data types to ones recognized by both linux and windows compilers

* added 'f' to float constants

* updated vulkan object initialization, windows does not allow .name listing in parameters until c++20

* updated to work on windows

* updated to work on windows

* updated to use wgl on windows

* updated to use wgl on windows

* exported object must be formally declared as such on windows

* added stub for setting an environment variable on windows

* vulkan flag bits must be formally cast on windows

* updated to build on windows

* updated build scripts for windows

* utils updates for windows

* updated src files for windows

* removed redundant include

* removed redundant find_package

* updated to run on windows

* needed for windows

* added new test profile for windows

* updated tcp connection to work on windows

* added because symlinks don't work on windows

* tcp networking test code

* tcp backend updated to work on windows

* fix for buggy windows library

* added no op record logger

* fixes for yaml-cpp

* added protobuf

* added missing header to cmake

* fix for relative clock type

* only output scripts on unix

* updated externalproject to use fetchcontent

* export the proper symbols for monado running

* updates to package requirements

* WIP

* first few conversions

* initial conversion of some external projects

* added ports and vcpkg files

* reordered to enable vcpkg use

* added checks for OS

* minor bug fixes for linux build

* added OS specific warnings

* removed un-needed packages

* removed test profiles

* fixed up package finding

* better fix for namespacing

* removed un-needed code

* make sure the library is properly found

* [pre-commit.ci] Run clang-format

* revert to external project

* fixed variable name

* added error checks

* revert to external_project

* fixed some typos from the conversion

* turn off dep_mapper for now

* update tags

* revert some dependencies to externalproject for now

* fix ffmpeg build

* fix vulkan build

* fix monado build

* added missing ifdef

* added casts

* fixed ffmpeg dependency

* fixed typos

* fixed header locations

* updated documentation for windows

* [pre-commit.ci] Run clang-format

* updated openvins

* added ifdef

* updated offline_cam for windows

* updated offline_imu for windows

* initial openwarp (not working yet)

* updated passthrough_integrator for windows

* updated rk4_integrator for windows

* added some windows configs for testing

* updated fauxpose for windows

* updated pose_prediction for windows

* set opencv to 4.5.5

* fixed timing issue

* force opencv version

* include fixes for windows

* added missing library

* fixes to reduce compiler warnings

* fixes for windows build

* minor cmake fixes

* added full versioning for vcpkg

* dependency fixes

* renamed vcpkg features to comply with naming standard

* WIP on windows_port_experimental

* fixed bad patches

* fixed include order for windows

* fixed incorrect function call

* added some compile flags for windows

* fixed function call for linux

* fixed include order

* added windows config files

* added logos denoting which plugins are supported on what platforms

* updates for config, build, and launch settings

* enforce the header import order

* [pre-commit.ci] Run clang-format

* removed old profiles

* fixed typo in name

* not ready for windows yet

* updated documentation generation

* updated windows based docs

* fixed typos

* updated windows build instructions

* bug fixes

* [pre-commit.ci] Run clang-format

* added PIC to vulkan utils build

* documentation update for conda

---------

+++ plugins/offline_imu/plugin.cpp
@@ -50,7 +50,7 @@ ILLIXR::threadloop::skip_option offline_imu::_p_should_skip() {
 
 void offline_imu::_p_one_iteration() {
     assert(sensor_data_it_ != sensor_data_.end());
-    time_point          real_now(std::chrono::duration<long, std::nano>{dataset_now_ - dataset_first_time_});
+    time_point          real_now(std::chrono::duration<long long, std::nano>{dataset_now_ - dataset_first_time_});
     const sensor_types& sensor_datum = sensor_data_it_->second;
 
     imu_.put(imu_.allocate<imu_type>(imu_type{real_now, (sensor_datum.imu0.angular_v), (sensor_datum.imu0.linear_a)}));
